+++ src/Faker/Internet.cs
@@ -68,7 +68,7 @@ namespace RimuTec.Faker {
       }
 
       /// <summary>
-      /// Generates an IP V4 address. Example: "24.29.18.175"
+      /// Generates an IP v4 address. Example: "24.29.18.175"
       /// </summary>
       /// <remarks>The method may return IP V4 addresses that may be reserved, private or public. If you need 
       /// a private address outside of any reserved range use <see cref="PrivateIPv4Address"/>. If you 
@@ -77,6 +77,24 @@ namespace RimuTec.Faker {
       public static string IPv4Address() {
          var ary = IntHelper.Repeat(0, 255, x => x).ToArray();
          return $"{ary.Sample()}.{ary.Sample()}.{ary.Sample()}.{ary.Sample()}";
+      }
+
+      /// <summary>
+      /// Returns an IP v4 address including mask. Example: "24.29.18.175/21"
+      /// </summary>
+      /// <returns></returns>
+      public static string IPv4CIDR() {
+         return $"{IPv4Address()}/{RandomNumber.Next(1, 31)}";
+      }
+
+      /// <summary>
+      /// Returns an IP v6 address. Example: "ac5f:d696:3807:1d72:2eb5:4e81:7d2b:e1df"
+      /// </summary>
+      /// <returns></returns>
+      public static string IPv6Address() {
+         var list = new List<string>();
+         8.TimesDo(x => list.Add($"{RandomNumber.Next(65536):x4}"));
+         return string.Join(":", list);
       }
 
       /// <summary>
@@ -123,7 +141,7 @@ namespace RimuTec.Faker {
       }
 
       /// <summary>
-      /// Generates a private IP V4 address. Example: "10.0.0.1"
+      /// Generates a private IP v4 address. Example: "10.0.0.1"
       /// </summary>
       /// <returns></returns>
       public static string PrivateIPv4Address() {
@@ -135,7 +153,7 @@ namespace RimuTec.Faker {
       }
 
       /// <summary>
-      /// Generates a public IP V4 address, guaranteed not to be in the IP range from <see cref="PrivateIPv4Address"/>. 
+      /// Generates a public IP v4 address, guaranteed not to be in the IP range from <see cref="PrivateIPv4Address"/>. 
       /// Example: "24.29.18.175"
       /// </summary>
       /// <returns></returns>